# api/openapi.yaml
openapi: "3.1.0"
info:
  title: FamilyRoots API
  version: "1.0.0"
  description: |
    ## FamilyRoots REST API — v1

    A multi-tenant genealogy platform API following REST conventions.

    ### Versioning
    URL-path versioning: all v1 endpoints are prefixed `/api/v1`.
    Breaking changes increment the major version (`/api/v2`).
    Minor/patch changes are backward-compatible and do not change the URL.

    ### Authentication
    Two tokens are issued on login:
    - **Access token** — short-lived JWT (15 min), passed as `Authorization: Bearer <token>`
    - **Refresh token** — long-lived (30 days), stored as an `HttpOnly SameSite=Strict` cookie

    ### Multi-tenancy
    Every request is scoped to a tenant derived from the JWT `tenant_id` claim.
    Tenant is never supplied by the caller — it is always resolved server-side from the token.

    ### Rate limiting
    All endpoints return rate-limit headers:
    - `X-RateLimit-Limit` — requests allowed in the current window
    - `X-RateLimit-Remaining` — requests remaining in the current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    - `Retry-After` — seconds to wait (only present on 429 responses)

    Default limits by plan:
    | Plan        | Requests/min | Burst |
    |-------------|-------------|-------|
    | FREE        | 60          | 10    |
    | BASIC       | 300         | 30    |
    | PREMIUM     | 1000        | 100   |
    | PROFESSIONAL| 5000        | 500   |

...
    ```
...
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued on login. Expires in 15 minutes.
        Payload includes: sub (user_id), tenant_id, email, roles, plan.

    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        HttpOnly SameSite=Strict cookie holding the refresh token.
        Used only on POST /auth/refresh and POST /auth/logout.

  # ──────────────────────────────────────────────────────────
  # COMMON PARAMETERS
  # ──────────────────────────────────────────────────────────
  parameters:
    treeId:
      name: treeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Family tree identifier
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: Maximum number of items to return
    cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque pagination cursor from a previous response

  # ──────────────────────────────────────────────────────────
  # COMMON HEADERS
  # ──────────────────────────────────────────────────────────
...
  schemas:
...
    PrivacyLevel:
      type: string
      enum: [PUBLIC, FAMILY, PRIVATE]
...
    CollaboratorRole:
      type: string
      enum: [VIEWER, CONTRIBUTOR, EDITOR, ADMIN]

    # ── Error ───────────────────────────────────────────────
    Error:
      type: object
      required: [type, title, status, requestId]
      properties:
        type:
          type: string
          format: uri
          example: "https://familyroots.app/errors/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: "The request body contains invalid fields."
        instance:
          type: string
          example: "/api/v1/trees/abc/persons"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        requestId:
          type: string
          example: "req_01hzab3k7x"
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
          example: "sex"
        code:
          type: string
          example: "INVALID_ENUM"
        message:
          type: string
          example: "Must be one of: MALE, FEMALE, OTHER, UNKNOWN"

    # ── Pagination ──────────────────────────────────────────
    Pagination:
      type: object
      required: [hasMore]
      properties:
        cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page (null if no more pages)
        hasMore:
          type: boolean
        totalCount:
          type: integer
          description: Total matching records (expensive; only returned when requested)

    # ── Auth / Tokens ───────────────────────────────────────
...
    Tree:
      type: object
      required: [id, name, ownerId, isPublic, privacyLevel, personCount, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        ownerId:
          type: string
          format: uuid
        isPublic:
          type: boolean
        privacyLevel:
          $ref: "#/components/schemas/PrivacyLevel"
        personCount:
          type: integer
        myRole:
          $ref: "#/components/schemas/CollaboratorRole"
          description: The calling user's role in this tree
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTreeRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
          nullable: true
        isPublic:
          type: boolean
          default: false
        privacyLevel:
          $ref: "#/components/schemas/PrivacyLevel"
          default: FAMILY
    UpdateTreeRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
          nullable: true
        isPublic:
          type: boolean
        privacyLevel:
          $ref: "#/components/schemas/PrivacyLevel"
...
  responses:
    400BadRequest:
      description: Bad Request — malformed JSON or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    401Unauthorized:
      description: Unauthorized — missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    403Forbidden:
      description: Forbidden — valid token but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    422UnprocessableEntity:
      description: Unprocessable Entity — validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    429TooManyRequests:
      description: Too Many Requests — rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# ──────────────────────────────────────────────────────────────
# GLOBAL SECURITY (override per-operation where needed)
# ──────────────────────────────────────────────────────────────
security:
  - BearerAuth: []

# ──────────────────────────────────────────────────────────────
# TAGS
# ──────────────────────────────────────────────────────────────
...
paths:
...
  /trees:
    get:
      tags: [Trees]
      operationId: listTrees
      summary: List trees the current user owns or collaborates on
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/CollaboratorRole"
          description: Filter by user's role in the tree
      responses:
        "200":
          description: List of trees
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tree"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/401Unauthorized"

    post:
      tags: [Trees]
      operationId: createTree
      summary: Create a new family tree
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTreeRequest"
      responses:
        "201":
          description: Tree created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tree"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "422":
          $ref: "#/components/responses/422UnprocessableEntity"
        "429":
          $ref: "#/components/responses/429TooManyRequests"
  /trees/{treeId}:
    parameters:
      - $ref: "#/components/parameters/treeId"
    get:
      tags: [Trees]
      operationId: getTree
      summary: Get a family tree by ID
      responses:
        "200":
          description: Tree detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tree"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"

    patch:
      tags: [Trees]
      operationId: updateTree
      summary: Update a tree's metadata
      description: Requires ADMIN or EDITOR role, or tree ownership
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTreeRequest"
      responses:
        "200":
          description: Updated tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tree"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"

    delete:
      tags: [Trees]
      operationId: deleteTree
      summary: Delete a tree and all its data
      description: Owner only. Irreversible after 30-day grace period.
      responses:
        "202":
          description: Tree deletion scheduled
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
...